[OPT]update to zy_compass_3.6:     1. refine operators: LRN, lstm, gridsample, gelu, silu and other acivation fuctions     2. add operators: atanh, affinegrid     3. add global calibration strategy: smooth_quant, awq, gptq     4. add passes for transformer: split_qkv_fc, detect_inf_mask_nodes, optimize_wdc     5. optimize insert cast op for lib     6. add debug api in optimizer_forward for co-debug with executor engine     7. bugs fix and codes refinement

+++ AIPUBuilder/Optimizer/ops/roialign.py
@@ -503,7 +503,8 @@ def roialign(self, *args):
                 split_rois = torch.split(rois, split_size_or_sections=1, dim=-1)
                 rois = torch.cat([split_rois[0], split_rois[2], split_rois[1], split_rois[4], split_rois[3]], dim=-1)
                 fm = nhwc2nchw(inp_d)
-                out = roi_align(fm, rois, [pool_height, pool_width], spatial_scale=spatial_x, sampling_ratio=sample_w)
+                out = roi_align(fm.float(), rois, [pool_height, pool_width],
+                                spatial_scale=spatial_x, sampling_ratio=sample_w)
                 out = nchw2nhwc(out)
         else:  # method=='max'
             params = {}
